… Lib/glyphsLib/data/GlyphData_Ideographs.xml
… Lib/glyphsLib/glyphdata_generated.py
… MetaTools/generate_glyphdata.py
… setup.py
… tests/glyphdata_test.py