Instead of making https://github.com/googlei18n/glyphsLib/pull/403 work, include GlyphData.xml and just parse it.