@madig 2018-09-05
Hm, could be something to do after we figure out the basics´.