@madig 2018-09-05
Note to self: make data dicts return empty dict so i don't have to do `or {}` everywhere.