@anthrotype 2018-09-05
> make data dicts return empty dict so i don't have to do or {} everywhere

just do `data.get(key, {})`