@madig 2018-09-06
I installed this branch in a fresh venv (no `-e`) and got
```
In [2]: %timeit %run -i "C:\\UsersLocal\\nikolaus.waxweiler\\Development\\tmp\\tet.py"
1.51 ms ± 12.3 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [3]: %timeit %run -i "C:\\UsersLocal\\nikolaus.waxweiler\\Development\\tmp\\tet.py"
1.5 ms ± 11.5 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```