@madig 2018-09-06
Including it yields for this branch:
```
In [1]: %timeit %run -i "C:\\UsersLocal\\nikolaus.waxweiler\\Development\\tmp\\tet.py"
1.58 ms ± 216 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

In [2]: %timeit %run -i "C:\\UsersLocal\\nikolaus.waxweiler\\Development\\tmp\\tet.py"
1.52 ms ± 13 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
So, slightly faster than baseline I suppose.