@anthrotype 2018-09-06
this PR would fix https://github.com/googlei18n/glyphsLib/issues/141