@anthrotype 2018-09-06
and would fix this as well https://github.com/googlei18n/glyphsLib/issues/402

shall we also remove that generate_glyphdata.py script?